cwh-0.238 forgot to rename default com port settings in config file. Add setter for json Take care of null settings Split tests to ensure we can execute non hardware tests on every build
