+++ host/src/config.properties
@@ -28,7 +28,7 @@ printFileProcessor.org.area515.resinprinter.text.TextFilePrintFileProcessor=true
 #https://github.com/area515/Creation-Workshop-Host/files/129785/RS232_codes.pdf
 #https://github.com/area515/Creation-Workshop-Host/files/88303/RS232_Control_Codes_H6510BD.pdf
 #http://www.viewsoniceurope.com/asset-files/files/user_guide/pjd7820hd/28077.pdf
-hexCodeBasedProjectors=[{"name":"Vivitek (D535)","comPortSettings":{"speed":9600, "databits":8, "parity":"None", "stopbits":"One"}, "onHex":"7E504E0D", "offHex":"7E50460D","bulbHoursHex":"7E714C0D","bulbHoursResponseRegex":"[^\\\\d]*(\\\\d+)\\\\x0D?", "bulbHoursConversion":"ASCII", "detectionHex":"7E504E0D(function (){java.lang.Thread.sleep(1);})()7E71560D", "detectionResponseRegex":"(?s:.*DP3406.*\\\\x0D?)"}, {"name":"Acer","comPortSettings":{"speed":9600, "databits":8, "parity":"None", "stopbits":"One"},"onHex":"4F4B4F4B4F4B4F4B4F4B0D", "offHex":"2A2030204952203030320D", "detectionHex":"2A2030204C616D70203F0D", "detectionResponseRegex":"(?s:.*?)(?i:Lamp) [10](?s:.*?)"}, {"name":"ViewSonic (pjd7820hd)","comPortSettings":{"speed":115200, "databits":8, "parity":"None", "stopbits":"One"},"onHex":"0614000400341100005D", "offHex":"0614000400341101005E", "bulbHoursHex":"0714000500340000150163", "bulbHoursResponseRegex":"\\\\x05\\\\x14\\\\x00\\\\x06\\\\x00\\\\x00\\\\x00([\\\\x00-\\\\xFF]{2})\\\\x00\\\\x00\\\\x1E", "bulbHoursConversion":"LittleEndian", "detectionHex":"071400050034000011005E", "detectionResponseRegex":"(?s:.*(?:\\\\x05\\\\x14\\\\x00\\\\x03\\\\x00\\\\x00\\\\x00[\\\\x00\\\\x01][\\\\x17\\\\x18]|\\\\x03\\\\x14\\\\x00\\\\x00\\\\x00\\\\x14)\\\\x0D?)"}]
+hexCodeBasedProjectors=[{"name":"Vivitek (D535)","defaultComPortSettings":{"speed":9600, "databits":8, "parity":"None", "stopbits":"One"}, "onHex":"7E504E0D", "offHex":"7E50460D","bulbHoursHex":"7E714C0D(function (){java.lang.Thread.sleep(100);})()","bulbHoursResponseRegex":"[^\\\\d]*(\\\\d+)\\\\x0D?", "bulbHoursConversion":"ASCII", "detectionHex":"7E504E0D(function (){java.lang.Thread.sleep(1);})()7E71560D", "detectionResponseRegex":"(?s:.*DP3406.*\\\\x0D?)"}, {"name":"Acer","defaultComPortSettings":{"speed":9600, "databits":8, "parity":"None", "stopbits":"One"},"onHex":"4F4B4F4B4F4B4F4B4F4B0D", "offHex":"2A2030204952203030320D", "detectionHex":"2A2030204C616D70203F0D", "detectionResponseRegex":"(?s:.*?)(?i:Lamp) [10](?s:.*?)"}, {"name":"ViewSonic (pjd7820hd)","defaultComPortSettings":{"speed":115200, "databits":8, "parity":"None", "stopbits":"One"},"onHex":"0614000400341100005D", "offHex":"0614000400341101005E", "bulbHoursHex":"0714000500340000150163", "bulbHoursResponseRegex":"\\\\x05\\\\x14\\\\x00\\\\x06\\\\x00\\\\x00\\\\x00([\\\\x00-\\\\xFF]{2})\\\\x00\\\\x00\\\\x1E", "bulbHoursConversion":"LittleEndian", "detectionHex":"071400050034000011005E", "detectionResponseRegex":"(?s:.*(?:\\\\x05\\\\x14\\\\x00\\\\x03\\\\x00\\\\x00\\\\x00[\\\\x00\\\\x01][\\\\x17\\\\x18]|\\\\x03\\\\x14\\\\x00\\\\x00\\\\x00\\\\x14)\\\\x0D?)"}]
 
 SerialCommunicationsImplementation=org.area515.resinprinter.serial.JSSCCommPort
 #3d firmware autodetection is unreliable on these implementations
