+++ host/tests/org/area515/resinprinter/test/FullTestSuite.java
@@ -5,6 +5,7 @@ import org.area515.resinprinter.inkdetection.visual.LineTest;
 import org.area515.resinprinter.inkdetection.visual.TestVisualPrintMaterialDetector;
 import org.area515.resinprinter.job.AbstractPrintFileProcessorTest;
 import org.area515.resinprinter.network.LinuxNetworkManagerTest;
+import org.area515.resinprinter.projector.HexCodeBasedProjectorTesting;
 import org.area515.resinprinter.services.TestScriptAndTemplating;
 import org.area515.util.IOUtilitiesTest;
 import org.junit.runner.RunWith;
@@ -19,6 +20,7 @@ import org.junit.runners.Suite;
 	LinuxNetworkManagerTest.class,
 	AbstractPrintFileProcessorTest.class,
 	TestScriptAndTemplating.class,
+	HexCodeBasedProjectorTesting.class,
 })
 
 public class FullTestSuite {
